move surveyChecker and remove unnused pullsdk classes
